refactor: new folder structure
