+++ .vscode/settings.json
@@ -6,16 +6,5 @@
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true,
     "explorer.excludeGitIgnore": false,
-    "files.exclude": {
-        "**/.git": true,
-        "**/.svn": true,
-        "**/.hg": true,
-        "**/CVS": true,
-        "**/.DS_Store": true,
-        "**/Thumbs.db": true,
-        "**/BUILD": true,
-        "*.egg-info": true,
-        "*.pytest_cache": true,
-    },
     "python.analysis.diagnosticMode": "workspace"
 }
