+++ personal_mnemonic_medium/data_access/exporters/anki/anki_css.py
