+++ personal_mnemonic_medium/data_access/exporters/anki/card_types/base.py
