+++ personal_mnemonic_medium/data_access/exporters/anki/globals.py
