+++ personal_mnemonic_medium/data_access/exporters/base.py
