+++ personal_mnemonic_medium/domain/card_pipeline.py
