+++ personal_mnemonic_medium/domain/prompt_extractors/base.py
