+++ personal_mnemonic_medium/domain/prompt_extractors/qa_extractor.py
