+++ personal_mnemonic_medium/domain/url_generators/obsidian_url.py
